examples/grpc_client/requirements.txt
numpy==1.15.2
Pillow==5.3.0
tensorflow-serving-api==1.11.0
